+++ src/Opulence/Ioc/Tests/Bootstrappers/Dispatchers/BootstrapperDispatcherTest.php
@@ -10,7 +10,6 @@
 
 namespace Opulence\Ioc\Tests\Bootstrappers\Dispatchers;
 
-use Opulence\Ioc\Bootstrappers\BootstrapperRegistry;
 use Opulence\Ioc\Bootstrappers\Dispatchers\BootstrapperDispatcher;
 use Opulence\Ioc\Container;
 use Opulence\Ioc\IContainer;
@@ -31,98 +30,54 @@ use Opulence\Ioc\Tests\Bootstrappers\Mocks\LazyFooInterface;
 class BootstrapperDispatcherTest extends \PHPUnit\Framework\TestCase
 {
     /** @var BootstrapperDispatcher The dispatcher to use in tests */
-    private $dispatcher = null;
+    private $dispatcher;
     /** @var Container The container to use in tests */
-    private $container = null;
-    /** @var BootstrapperRegistry The bootstrapper registry */
-    private $bootstrapperRegistry = null;
+    private $container;
 
-    /**
-     * Sets up the tests
-     */
     public function setUp() : void
     {
         $this->container = new Container();
-        $this->bootstrapperRegistry = new BootstrapperRegistry();
-        $this->dispatcher = new BootstrapperDispatcher(
-            $this->container,
-            $this->bootstrapperRegistry
-        );
+        $this->dispatcher = new BootstrapperDispatcher($this->container);
     }
 
-    /**
-     * Tests dispatching all bootstrappers eagerly
-     */
     public function testDispatchingAllBootstrappersEagerly() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(new EagerBootstrapper());
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
-        $this->dispatcher->dispatch(true);
+        $this->dispatcher->dispatch([new EagerBootstrapper(), new LazyBootstrapper()]);
         $this->assertInstanceOf(LazyConcreteFoo::class, $this->container->resolve(LazyFooInterface::class));
         $this->assertInstanceOf(EagerConcreteFoo::class, $this->container->resolve(EagerFooInterface::class));
     }
 
-    /**
-     * Tests dispatching an eager bootstrapper that depends on a dependency set in a lazy bootstrapper
-     */
-    public function testDispatchingEagerDispatcherThatDependsOnDependencyFromLazyBootstrapper() : void
-    {
-        $this->bootstrapperRegistry->registerBootstrapper(new EagerBootstrapperThatDependsOnBindingFromLazyBootstrapper());
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
-        ob_start();
-        $this->dispatcher->dispatch(false);
-        $this->assertEquals(LazyConcreteFoo::class, ob_get_clean());
-    }
-
-    /**
-     * Tests dispatching a lazy bootstrapper (registered first) that depends on a dependency set in a lazy bootstrapper
-     */
     public function testDispatchingLazyDispatcherThatDependsOnDependencyFromLazyBootstrapperAndRegisteringItFirst() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(
-            new LazyBootstrapperThatDependsOnBindingFromLazyBootstrapper()
-        );
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
         ob_start();
-        $this->dispatcher->dispatch(false);
+        $this->dispatcher->dispatch([
+            new LazyBootstrapperThatDependsOnBindingFromLazyBootstrapper(),
+            new LazyBootstrapper()
+        ]);
         $this->container->resolve(EagerFooInterface::class);
         $this->assertEquals(LazyConcreteFoo::class, ob_get_clean());
     }
 
-    /**
-     * Tests dispatching a lazy bootstrapper (registered second) that depends on a dependency set in a lazy bootstrapper
-     */
     public function testDispatchingLazyDispatcherThatDependsOnDependencyFromLazyBootstrapperAndRegisteringItSecond() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
-        $this->bootstrapperRegistry->registerBootstrapper(
-            new LazyBootstrapperThatDependsOnBindingFromLazyBootstrapper()
-        );
         ob_start();
-        $this->dispatcher->dispatch(false);
+        $this->dispatcher->dispatch([
+            new LazyBootstrapper(),
+            new LazyBootstrapperThatDependsOnBindingFromLazyBootstrapper()
+        ]);
         $this->container->resolve(EagerFooInterface::class);
         $this->assertEquals(LazyConcreteFoo::class, ob_get_clean());
     }
 
-    /**
-     * Tests that a lazy bootstrapper's bindings are available to the container
-     */
     public function testLazyBootstrappersBindingsAreAvailableToContainer() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
-        $this->dispatcher->dispatch(false);
+        $this->dispatcher->dispatch([new LazyBootstrapper()]);
         $this->assertInstanceOf(LazyConcreteFoo::class, $this->container->resolve(LazyFooInterface::class));
     }
 
-    /**
-     * Tests that lazy targeted bindings are available in the container
-     */
     public function testLazyTargetedBindingsAreAvailableInContainer() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(
-            new LazyBootstrapperWithTargetedBinding()
-        );
-        $this->dispatcher->dispatch(false);
+        $this->dispatcher->dispatch([new LazyBootstrapperWithTargetedBinding()]);
         // Make sure it's bound to the target
         $this->assertTrue(
             $this->container->for(
@@ -155,14 +110,9 @@ class BootstrapperDispatcherTest extends \PHPUnit\Framework\TestCase
         $this->assertFalse($this->container->hasBinding(LazyFooInterface::class));
     }
 
-    /**
-     * Tests not dispatching all bootstrappers eagerly
-     */
     public function testNotDispatchingAllBootstrappersEagerly() : void
     {
-        $this->bootstrapperRegistry->registerBootstrapper(new EagerBootstrapper());
-        $this->bootstrapperRegistry->registerBootstrapper(new LazyBootstrapper());
-        $this->dispatcher->dispatch(false);
+        $this->dispatcher->dispatch([new EagerBootstrapper(), new LazyBootstrapper()]);
         $this->assertTrue($this->container->hasBinding(LazyFooInterface::class));
         $this->assertInstanceOf(EagerConcreteFoo::class, $this->container->resolve(EagerFooInterface::class));
     }