+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * Added `IContainer::tryResolve()` to simplify trying to resolve something without needing `try`/`catch` blocks
 * Made `Opulence\Ioc\Bootstrappers\Bootstrapper` abstract, added final constructor
 * Removed `ILazyBootstrapper` and made a new, abstract `LazyBootstrapper` to be used instead
+* Changed `BootstrapperDispatcher::dispatch()` to take in an array of instantiated bootstrappers rather than having a dependency on `IBootstrapperRegistry`
 * Removed `IBootstrapperResolver` and `BootstrapperResolver`
 
 <h2>v1.1.3 (2017-12-22)</h2>